# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/webhooks": {
      "post": {
        "operationId": "WebhooksController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebhookEndpointDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookEndpointWithSecretEntity"
                }
              }
            }
          }
        },
        "summary": "Register a webhook endpoint (returns the signing secret — shown once)",
        "tags": [
          "webhooks"
        ]
      },
      "get": {
        "operationId": "WebhooksController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebhookEndpointEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "List the consumer's webhook endpoints",
        "tags": [
          "webhooks"
        ]
      }
    },
    "/v1/webhooks/{id}": {
      "get": {
        "operationId": "WebhooksController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookEndpointEntity"
                }
              }
            }
          }
        },
        "summary": "Get a webhook endpoint",
        "tags": [
          "webhooks"
        ]
      },
      "patch": {
        "operationId": "WebhooksController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWebhookEndpointDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookEndpointEntity"
                }
              }
            }
          }
        },
        "summary": "Update a webhook endpoint (url/description/enabled/eventTypes)",
        "tags": [
          "webhooks"
        ]
      },
      "delete": {
        "operationId": "WebhooksController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookDeletedEntity"
                }
              }
            }
          }
        },
        "summary": "Delete a webhook endpoint",
        "tags": [
          "webhooks"
        ]
      }
    },
    "/v1/webhooks/{id}/rotate-secret": {
      "post": {
        "operationId": "WebhooksController_rotateSecret_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RotateWebhookSecretDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookEndpointWithSecretEntity"
                }
              }
            }
          }
        },
        "summary": "Rotate the signing secret (returns the new secret). The previous secret keeps signing deliveries until previousSecretExpiresAt, unless graceSeconds=0.",
        "tags": [
          "webhooks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Cosmos Pay — Payments API",
    "description": "Payments microservice (Stellar payment intents). All endpoints require traffic to arrive through the APISIX gateway: a valid `X-Gateway-Secret` header plus an authenticated consumer (`X-Consumer-Username`). Paths already include the version (`/v1/...`).",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "gateway-secret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Gateway-Secret"
      },
      "consumer": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Consumer-Username"
      }
    },
    "schemas": {
...
      "CreateWebhookEndpointDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "HTTPS URL that will receive POSTed event notifications.",
            "example": "https://integrator.example.com/webhooks/cosmos"
          },
          "description": {
            "type": "string",
            "example": "Production payment events"
          },
          "eventTypes": {
            "type": "array",
            "description": "Event types to subscribe to. Omit/empty to receive all events.",
            "items": {
              "type": "string",
              "enum": [
                "PAYMENT_INTENT_CREATED",
                "PAYMENT_INTENT_UPDATED",
                "PAYMENT_INTENT_SUCCEEDED",
                "PAYMENT_INTENT_FAILED",
                "PAYMENT_INTENT_CANCELLED",
                "PAYMENT_INTENT_DELETED",
                "RECEIVER_UPDATED",
                "PAYIN_CREATED",
                "PAYIN_UPDATED",
                "PAYIN_COMPLETED",
                "PAYOUT_CREATED",
                "PAYOUT_UPDATED",
                "PAYOUT_COMPLETED",
                "SWAP_CREATED",
                "SWAP_SUBMITTED",
                "SWAP_SUCCEEDED",
                "SWAP_FAILED",
                "SWAP_EXPIRED",
                "LIQUIDITY_CREATED",
                "LIQUIDITY_SUBMITTED",
                "LIQUIDITY_SUCCEEDED",
                "LIQUIDITY_FAILED",
                "LIQUIDITY_EXPIRED"
              ]
            }
          }
        },
        "required": [
          "url"
        ]
      },
      "WebhookEndpointWithSecretEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "we_9z8a1b0000abcd1234efgh"
          },
          "url": {
            "type": "string",
            "example": "https://integrator.example.com/webhooks/cosmos"
          },
          "description": {
            "type": "object",
            "nullable": true,
            "example": "Production payment events"
          },
          "enabled": {
            "type": "boolean",
            "example": true
          },
          "destinationBlocked": {
            "type": "boolean",
            "example": false,
            "description": "True when the URL failed public-destination checks (SSRF policy). Deliveries are skipped until the URL is updated to a safe https endpoint."
          },
          "eventTypes": {
            "type": "array",
            "example": [
              "PAYMENT_INTENT_SUCCEEDED"
            ],
            "description": "Subscribed events; empty means all.",
            "items": {
              "type": "string",
              "enum": [
                "PAYMENT_INTENT_CREATED",
                "PAYMENT_INTENT_UPDATED",
                "PAYMENT_INTENT_SUCCEEDED",
                "PAYMENT_INTENT_FAILED",
                "PAYMENT_INTENT_CANCELLED",
                "PAYMENT_INTENT_DELETED",
                "RECEIVER_UPDATED",
                "PAYIN_CREATED",
                "PAYIN_UPDATED",
                "PAYIN_COMPLETED",
                "PAYOUT_CREATED",
                "PAYOUT_UPDATED",
                "PAYOUT_COMPLETED",
                "SWAP_CREATED",
                "SWAP_SUBMITTED",
                "SWAP_SUCCEEDED",
                "SWAP_FAILED",
                "SWAP_EXPIRED",
                "LIQUIDITY_CREATED",
                "LIQUIDITY_SUBMITTED",
                "LIQUIDITY_SUCCEEDED",
                "LIQUIDITY_FAILED",
                "LIQUIDITY_EXPIRED"
              ]
            }
          },
          "previousSecretExpiresAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "example": "2026-08-26T17:00:00.000Z",
            "description": "When the previous signing secret stops being accepted on deliveries. Null when no rotation is in flight. The previous secret value itself is never returned."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-06-21T12:34:56.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-06-21T12:34:56.000Z"
          },
          "secret": {
            "type": "string",
            "description": "HMAC signing secret — shown once. Store it securely.",
            "example": "whsec_3f9a1c…"
          }
        },
        "required": [
          "id",
          "url",
          "description",
          "enabled",
          "destinationBlocked",
          "eventTypes",
          "previousSecretExpiresAt",
          "createdAt",
          "updatedAt",
          "secret"
        ]
      },
      "WebhookEndpointEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "we_9z8a1b0000abcd1234efgh"
          },
          "url": {
            "type": "string",
            "example": "https://integrator.example.com/webhooks/cosmos"
          },
          "description": {
            "type": "object",
            "nullable": true,
            "example": "Production payment events"
          },
          "enabled": {
            "type": "boolean",
            "example": true
          },
          "destinationBlocked": {
            "type": "boolean",
            "example": false,
            "description": "True when the URL failed public-destination checks (SSRF policy). Deliveries are skipped until the URL is updated to a safe https endpoint."
          },
          "eventTypes": {
            "type": "array",
            "example": [
              "PAYMENT_INTENT_SUCCEEDED"
            ],
            "description": "Subscribed events; empty means all.",
            "items": {
              "type": "string",
              "enum": [
                "PAYMENT_INTENT_CREATED",
                "PAYMENT_INTENT_UPDATED",
                "PAYMENT_INTENT_SUCCEEDED",
                "PAYMENT_INTENT_FAILED",
                "PAYMENT_INTENT_CANCELLED",
                "PAYMENT_INTENT_DELETED",
                "RECEIVER_UPDATED",
                "PAYIN_CREATED",
                "PAYIN_UPDATED",
                "PAYIN_COMPLETED",
                "PAYOUT_CREATED",
                "PAYOUT_UPDATED",
                "PAYOUT_COMPLETED",
                "SWAP_CREATED",
                "SWAP_SUBMITTED",
                "SWAP_SUCCEEDED",
                "SWAP_FAILED",
                "SWAP_EXPIRED",
                "LIQUIDITY_CREATED",
                "LIQUIDITY_SUBMITTED",
                "LIQUIDITY_SUCCEEDED",
                "LIQUIDITY_FAILED",
                "LIQUIDITY_EXPIRED"
              ]
            }
          },
          "previousSecretExpiresAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "example": "2026-08-26T17:00:00.000Z",
            "description": "When the previous signing secret stops being accepted on deliveries. Null when no rotation is in flight. The previous secret value itself is never returned."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-06-21T12:34:56.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-06-21T12:34:56.000Z"
          }
        },
        "required": [
          "id",
          "url",
          "description",
          "enabled",
          "destinationBlocked",
          "eventTypes",
          "previousSecretExpiresAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateWebhookEndpointDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://integrator.example.com/webhooks/cosmos"
          },
          "description": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean",
            "description": "Pause/resume deliveries to this endpoint."
          },
          "eventTypes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "PAYMENT_INTENT_CREATED",
                "PAYMENT_INTENT_UPDATED",
                "PAYMENT_INTENT_SUCCEEDED",
                "PAYMENT_INTENT_FAILED",
                "PAYMENT_INTENT_CANCELLED",
                "PAYMENT_INTENT_DELETED",
                "RECEIVER_UPDATED",
                "PAYIN_CREATED",
                "PAYIN_UPDATED",
                "PAYIN_COMPLETED",
                "PAYOUT_CREATED",
                "PAYOUT_UPDATED",
                "PAYOUT_COMPLETED",
                "SWAP_CREATED",
                "SWAP_SUBMITTED",
                "SWAP_SUCCEEDED",
                "SWAP_FAILED",
                "SWAP_EXPIRED",
                "LIQUIDITY_CREATED",
                "LIQUIDITY_SUBMITTED",
                "LIQUIDITY_SUCCEEDED",
                "LIQUIDITY_FAILED",
                "LIQUIDITY_EXPIRED"
              ]
            }
          }
        }
      },
      "WebhookDeletedEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "we_9z8a1b0000abcd1234efgh"
          },
          "deleted": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "id",
          "deleted"
        ]
      },
      "RotateWebhookSecretDto": {
        "type": "object",
        "properties": {
          "graceSeconds": {
            "type": "number",
            "description": "Overlap window in seconds during which deliveries are signed with both the new and previous secrets. Defaults to WEBHOOK_SECRET_GRACE_SECONDS. Pass 0 to revoke the previous secret immediately (confirmed leak). Cannot exceed the configured maximum.",
            "example": 86400,
            "minimum": 0
          }
        }
      },
...
    }
  },
  "security": [
    {
      "gateway-secret": []
    },
    {
      "consumer": []
    }
  ]
}